Cain & Abel 4.9.56 @ https://gundam.technology
Cain and Abel (often abbreviated to Cain) is a password recovery tool for Microsoft Windows. It can recover many kinds of passwords using methods such as network packet sniffing, cracking various password hashes by using methods such as dictionary attacks, brute force and cryptanalysis attacks.[1] Cryptanalysis attacks are done via rainbow tables which can be generated with the winrtgen.exe program provided with Cain and Abel.

Features

    WEP cracking
    Speeding up packet capture speed by wireless packet injection
    Ability to record VoIP conversations
    Decoding scrambled passwords
    Calculating hashes
    Traceroute
    Revealing password boxes
    Uncovering cached passwords
    Dumping protected storage passwords
    ARP spoofing
    IP to MAC Address resolver
    Network Password Sniffer
    LSA secret dumper
    Ability to crack:
        LM & NTLM hashes
        NTLMv2 hashes
        Microsoft Cache hashes
        Microsoft Windows PWL files
        Cisco IOS - MD5 hashes
        Cisco PIX - MD5 hashes
        APOP - MD5 hashes
        CRAM-MD5 MD5 hashes
        OSPF - MD5 hashes
        RIPv2 MD5 hashes
        VRRP - HMAC hashes
        Virtual Network Computing (VNC) Triple DES
        MD2 hashes
        MD4 hashes
        MD5 hashes
        SHA-1 hashes
        SHA-2 hashes
        RIPEMD-160 hashes
        Kerberos 5 hashes
        RADIUS shared key hashes
        IKE PSK hashes
        MSSQL hashes
        MySQL hashes
        Oracle and SIP hashes

Status with virus scanners

Some virus scanners (and browsers, e.g. Google Chrome 20.0.1132.47) detect Cain and Abel as malware.

Avast! detects it as "Win32:Cain-B [Tool]" and classifies it as "Other potentially dangerous program", while Microsoft Security Essentials detects it as "Win32/Cain!4_9_14" and classifies it as "Tool: This program has potentially unwanted behavior." Even if Cain's install directory, as well as the word "Cain", are added to Avast's exclude list, the real-time scanner has been known to stop Cain from functioning. However, the latest version of Avast no longer blocks Cain.

Symantec (the developer of the Norton family of computer security software) identified a buffer overflow vulnerability in version 4.9.24 that allowed for remote code execution in the event the application was used to open a large RDP file, as might occur when using the program to analyze network traffic. The vulnerability had been present in the previous version (4.9.23) as well and was patched in a subsequent release.